fix: errors about not found exception restaurants

+++ src/modules/restaurants/restaurants.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateRestaurantDto } from './dto/create-restaurant.dto';
 import { RestaurantsRepository } from './repositories/restaurants.repository';
 import { UpdateRestaurantDto } from './dto/update-restaurant.dto';
@@ -20,10 +20,21 @@ export class RestaurantsService {
 
   async findOne(id: string) {
     const restaurant = await this.restaurantRepository.findOne(id);
+
+    if (!restaurant) {
+      throw new NotFoundException('This restaurant does not exist');
+    }
+
     return restaurant;
   }
 
   async update(id: string, updateRestaurantDto: UpdateRestaurantDto) {
+    const restaurant = await this.restaurantRepository.findOne(id);
+
+    if (!restaurant) {
+      throw new NotFoundException('This restaurant does not exist');
+    }
+
     const updatedRestaurant = await this.restaurantRepository.update(
       id,
       updateRestaurantDto,
@@ -33,6 +44,11 @@ export class RestaurantsService {
   }
 
   async remove(id: string) {
+    const restaurant = await this.restaurantRepository.findOne(id);
+
+    if (!restaurant) {
+      throw new NotFoundException('This restaurant does not exist');
+    }
     return await this.restaurantRepository.delete(id);
   }
 }
